Simple_Agent_using_UV_OpenAI_Agents_SDK_and_Gemini/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]
